Extract default context
