+++ src/components/region-editor/RegionEditor.tsx
@@ -3,6 +3,7 @@ import snapshotImage from "../../assets/snapshot.png";
 import useLocalState from "../../hooks/UseLocalState.tsx";
 import { Rectangles } from "../shapes/Rectangle.tsx";
 import { useState } from "react";
+import { Box, Stack } from "@mui/material";
 
 export default function RegionEditor() {
   const [rectangles, setRectangles] = useLocalState<Rectangles>({}, "regions");
@@ -11,7 +12,7 @@ export default function RegionEditor() {
   );
 
   return (
-    <>
+    <Stack height="100%" justifyContent="center" alignItems="center">
       <ImageLabeller
         image={snapshotImage}
         rectangles={rectangles}
@@ -19,11 +20,11 @@ export default function RegionEditor() {
         selectedRectangleId={selectedRectangleId}
         setSelectedRectangleId={setSelectedRectangleId}
       />
-      <div>
+      <Box>
         Selected rectangle: {selectedRectangleId}{" "}
         {selectedRectangleId &&
           `: ${JSON.stringify(rectangles[selectedRectangleId])}`}
-      </div>
-    </>
+      </Box>
+    </Stack>
   );
 }
